[Refactor] include the GeoCLI class in the exports
